DeepAC-main/requirements.txt
pytorch-lightning==2.3.0
imgaug==0.4.0
opencv-python==4.6.0.66
pytorch3d==0.7.6
hydra-core==1.1.1
omegaconf==2.1.2
ray==2.1.0
timm==0.9.2
tqdm==4.64.1
segmentation-models-pytorch==0.3.3
coremltools==6.1
tensorboard==2.11.0
numpy==1.23.5
# imageio==2.22.4   
# openexr==1.3.9   
# pyexr==0.3.10     
